If a triangle is fit in a circle, then find the left area.
If your math problem asks you to find the remaining area after fitting a triangle in a circle, this source code is gonna make your life easier.
For example:
Q: A triangle whose base length is 4cm and height is 6 cm is fitted in a circle of radius 10cm. Then find out the remaining area in the circle. (Hint:Use this code).